refactor: 로또 생성 기능 (자동, 수동) 서비스로 변경
